Fix `thread::spawn` 'static lifetime issue.
This commit makes wamp-rs compile without boxed callbacks but the
implementation is not practical as the use of closures is mostly prohibited
be the fact that the callback has to at least live as long as the client
instance. The examples are currently failing as especially the pubsubclient
needs closures to work. Closures are generally living in the scope where the
register/subscribe method is called, which is most likely after the client
was instanciated (smaller lifetime than client -> compiler error).

This commit should not be considered as a solution!